+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add admin upload section for Supply Teachers ([PR 2404](https://github.com/Crown-Commercial-Service/crown-marketplace-legacy/pull/2404))
 - Add the supplier search for buyers ([PR 2405](https://github.com/Crown-Commercial-Service/crown-marketplace-legacy/pull/2405))
 - Update the title Supply Teachers RM6376 framework ([PR 2420](https://github.com/Crown-Commercial-Service/crown-marketplace-legacy/pull/2420))
+- Add ability for admin to update branches ([PR 2440](https://github.com/Crown-Commercial-Service/crown-marketplace-legacy/pull/2440))
 
 ### Changed